# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/courseelement": {
            "get": {
                "tags": [
                    "CourseElement"
                ],
                "summary": "Get list of CourseElement",
                "description": "Returns list of CourseElement",
                "operationId": "getCourseElement",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "CourseElement"
                ],
                "summary": "store CourseElement information",
                "description": "store CourseElement  data",
                "operationId": "storeCourseElement",
                "parameters": [
                    {
                        "name": "course_template_id",
                        "in": "path",
                        "description": "CourseElement course_template_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "CourseElement name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "path",
                        "description": "CourseElement StartDate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "applyTax",
                        "in": "path",
                        "description": "CourseElement EndDate",
                        "required": true,
                        "schema": {
                            "type": "bool"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/courseelement/{id}": {
            "get": {
                "tags": [
                    "CourseElement"
                ],
                "summary": "Get CourseElement information",
                "description": "Returns CourseElement data",
                "operationId": "getCourseElementById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CourseElement id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "CourseElement"
                ],
                "summary": "update CourseElement information",
                "description": "update CourseElement  data",
                "operationId": "updateCourseElement",
                "parameters": [
                    {
                        "name": "course_template_id",
                        "in": "path",
                        "description": "CourseElement course_template_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "CourseElement name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "path",
                        "description": "CourseElement StartDate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "applyTax",
                        "in": "path",
                        "description": "CourseElement EndDate",
                        "required": true,
                        "schema": {
                            "type": "bool"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CourseElement"
                ],
                "summary": "Delete CourseElement information",
                "description": "Delete CourseElement data",
                "operationId": "deleteCourseElementById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CourseElement id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/courseelement/restore/{id}": {
            "get": {
                "tags": [
                    "CourseElement"
                ],
                "summary": "Restore deleted CourseElement information",
                "description": "Restore deleted CourseElement data",
                "operationId": "restoreCourseElementById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CourseElement id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/courseelement/restore": {
            "get": {
                "tags": [
                    "CourseElement"
                ],
                "summary": "restore all deleted CourseElement information",
                "description": "restore all  CourseElement data",
                "operationId": "restoreAllCourseElement",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
    },
...
}